@@ -14,8 +14,11 @@
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.4",
     "css-loader": "^3.4.2",
+    "node-sass": "^5.0.0",
+    "sass": "^1.30.0",
+    "sass-loader": "^10.1.0",
     "style-loader": "^1.1.2",
-    "webpack": "^4.20.2",
+    "webpack": "^4.44.2",
     "webpack-cli": "^3.1.2"
   },
   "dependencies": {
